+++ indigo/plugins.py
@@ -18,7 +18,8 @@ class LocaleBasedRegistry(object):
     def for_document(self, topic, document):
         """ Find an appropriate helper for this document.
         """
-        return self.for_locale(topic, country=document.country, locality=document.locality, language=document.language.code)
+        locality = document.work.locality.code if document.work.locality else None
+        return self.for_locale(topic, country=document.work.country.code, locality=locality, language=document.language.code)
 
     def for_work(self, topic, work):
         """ Find an appropriate helper for this work.