+++ indigo/settings.py
@@ -129,7 +129,7 @@ INDIGO = {
 
 import dj_database_url
 db_config = dj_database_url.config(default='postgres://indigo:indigo@localhost:5432/indigo')
-db_config['ATOMIC_REQUESTS'] = False
+db_config['ATOMIC_REQUESTS'] = True
 DATABASES = {
     'default': db_config,
 }
@@ -399,6 +399,9 @@ LOGGING = {
             'level': 'DEBUG' if DEBUG else 'INFO',
         },
         'indigo_app': {
+            'level': 'DEBUG' if DEBUG else 'INFO',
+        },
+        'indigo_metrics': {
             'level': 'DEBUG' if DEBUG else 'INFO',
         },
         'django': {