+++ indigo_api/templatetags/indigo.py
@@ -13,4 +13,9 @@ def work_resolver_url(context, work):
         frbr_uri = work.frbr_uri
     else:
         frbr_uri = work
+
+    if context.get('media_resolver_use_akn_prefix') and not frbr_uri.startswith('/akn'):
+        # for V2 APIs, prefix FRBR URI with /akn
+        frbr_uri = '/akn' + frbr_uri
+
     return context.get('resolver_url', settings.RESOLVER_URL) + frbr_uri