+++ indigo_app/static/javascript/indigo/views/document.js
@@ -183,6 +183,11 @@
       this.annotationsView.listenTo(this.bodyEditorView.sourceEditor, 'rendered', this.annotationsView.renderAnnotations);
 
       this.activityView = new Indigo.DocumentActivityView({document: this.document});
+      this.issuesView = new Indigo.DocumentIssuesView({
+        document: this.document,
+        documentContent: this.documentContent,
+        editorView: this.bodyEditorView,
+      });
 
       // pretend we've fetched it, this sets up additional handlers
       this.document.trigger('sync');