+++ indigo_app/views/documents.py
@@ -40,6 +40,7 @@ class DocumentDetailView(AbstractAuthedIndigoView, DetailView):
                 doc.work.expressions().all()
             ))
         context['comparison_expressions'] = doc.work.expressions().filter(language=doc.language).order_by('-expression_date')
+        context['place'] = doc.work.place
         context['country'] = doc.work.country
         context['locality'] = doc.work.locality